style(deploy)：一键部署命令优化，支持传入需要部署项目名

+++ deploy.sh
@@ -12,39 +12,47 @@ git submodule update --init --recursive
 git submodule update
 echo -e "\033[32m代码拉取完成...\033[0m"
 
-echo "开始部署后台管理(manager-admin)..."
-cd $WEB_PATH/manager-admin
-rm -rf node_modules
-echo "开始安装项目依赖..."
-sudo npm install
-echo "开始打包..."
-sudo npm run build:prod
-echo -e "\033[32m后台管理部署完成...\033[0m"
+if [[ ! -n "$1" || "$1" == "manager-admin" ]];then
+  echo "开始部署后台管理(manager-admin)..."
+  cd $WEB_PATH/manager-admin
+  rm -rf node_modules
+  echo "开始安装项目依赖..."
+  sudo npm install
+  echo "开始打包..."
+  sudo npm run build:prod
+  echo -e "\033[32m后台管理部署完成...\033[0m"
+fi
 
-echo "开始部署商家中心(manager-seller)..."
-cd $WEB_PATH/manager-seller
-rm -rf node_modules
-echo "开始安装项目依赖..."
-sudo npm install
-echo "开始打包..."
-sudo npm run build:prod
-echo -e "\033[32m商家中心部署完成...\033[0m"
+if [[ ! -n "$1" || "$1" == "manager-seller" ]];then
+  echo "开始部署商家中心(manager-seller)..."
+  cd $WEB_PATH/manager-seller
+  rm -rf node_modules
+  echo "开始安装项目依赖..."
+  sudo npm install
+  echo "开始打包..."
+  sudo npm run build:prod
+  echo -e "\033[32m商家中心部署完成...\033[0m"
+fi
 
-echo "开始部署买家PC端(buyer-pc)..."
-cd $WEB_PATH/buyer/pc/themes/b2b2c_v5
-rm -rf node_modules
-echo "开始安装项目依赖..."
-sudo npm install
-sudo npm run build
-echo -e "\033[32m买家PC端build完成，等待部署...\033[0m"
+if [[ ! -n "$1" || "$1" == "buyer-pc" ]];then
+  echo "开始部署买家PC端(buyer-pc)..."
+  cd $WEB_PATH/buyer/pc/themes/b2b2c_v5
+  rm -rf node_modules
+  echo "开始安装项目依赖..."
+  sudo npm install
+  sudo npm run build
+  echo -e "\033[32m买家PC端build完成，等待部署...\033[0m"
+fi
 
-# echo "开始部署买家WAP端(buyer-wap)..."
-# cd $WEB_PATH/buyer/wap/themes/default
-# rm -rf node_modules
-# echo "开始安装项目依赖..."
-# sudo npm install
-# sudo npm run build
-# echo "买家WAP端build完成，等待部署..."
+if [[ ! -n "$1" || "$1" == "buyer-wap" ]];then
+  #echo "开始部署买家WAP端(buyer-wap)..."
+  #cd $WEB_PATH/buyer/wap/themes/default
+  #rm -rf node_modules
+  #echo "开始安装项目依赖..."
+  #sudo npm install
+  #sudo npm run build
+  #echo "买家WAP端build完成，等待部署..."
+fi
 
 # 删除所有应用
 pm2 delete all
